[refactor] error관련 파일을 error 폴더로 이동
